backend/pyproject.toml
[tool.poetry]
name = "gt"
version = "0.1.0"
description = "challenge"
authors = ["Spadavecchia Adrian <spadavecchia.adrian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
fastapi = "^0.109.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
uvicorn = "^0.27.0"
SQLAlchemy = "^2.0.25"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
